chtml/html/testdata/webkit/tests25.dat

#data
<!DOCTYPE html><body><foo>A
#errors
(1,27): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <body>
|   <foo>
|     "A"

#data
<!DOCTYPE html><body><area>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <area>
|   "A"

#data
<!DOCTYPE html><body><base>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <base>
|   "A"

#data
<!DOCTYPE html><body><basefont>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <basefont>
|   "A"

#data
<!DOCTYPE html><body><bgsound>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <bgsound>
|   "A"

#data
<!DOCTYPE html><body><br>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <br>
|   "A"

#data
<!DOCTYPE html><body><col>A
#errors
(1,26): unexpected-start-tag-ignored
#document
| <!DOCTYPE html>
| <body>
|   <col>
|     "A"

#data
<!DOCTYPE html><body><command>A
#errors
eof
#document
| <!DOCTYPE html>
| <body>
|   <command>
|     "A"

#data
<!DOCTYPE html><body><embed>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <embed>
|   "A"

#data
<!DOCTYPE html><body><frame>A
#errors
(1,28): unexpected-start-tag-ignored
#document
| <!DOCTYPE html>
| <body>
|   <frame>
|     "A"

#data
<!DOCTYPE html><body><hr>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <hr>
|   "A"

#data
<!DOCTYPE html><body><img>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <img>
|   "A"

#data
<!DOCTYPE html><body><input>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <input>
|   "A"

#data
<!DOCTYPE html><body><keygen>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <keygen>
|   "A"

#data
<!DOCTYPE html><keygen>A</keygen>B
#errors
33: Stray end tag “keygen”.
#document
| <!DOCTYPE html>
| <keygen>
| "AB"

#data
</keygen>A
#errors
9: End tag seen without seeing a doctype first. Expected “<!DOCTYPE html>”.
9: Stray end tag “keygen”.
#document
| "A"

#data
<!DOCTYPE html></keygen>A
#errors
24: Stray end tag “keygen”.
#document
| <!DOCTYPE html>
| "A"

#data
<!DOCTYPE html><head></keygen>A
#errors
30: Stray end tag “keygen”.
#document
| <!DOCTYPE html>
| <head>
|   "A"

#data
<!DOCTYPE html><head></head></keygen>A
#errors
30: Stray end tag “keygen”.
#document
| <!DOCTYPE html>
| <head>
| "A"

#data
<!DOCTYPE html><body></keygen>A
#errors
30: Stray end tag “keygen”.
#document
| <!DOCTYPE html>
| <body>
|   "A"

#data
<!DOCTYPE html><body><link>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <link>
|   "A"

#data
<!DOCTYPE html><body><meta>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <meta>
|   "A"

#data
<!DOCTYPE html><body><param>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <param>
|   "A"

#data
<!DOCTYPE html><body><source>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <source>
|   "A"

#data
<!DOCTYPE html><body><track>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <track>
|   "A"

#data
<!DOCTYPE html><body><wbr>A
#errors
#document
| <!DOCTYPE html>
| <body>
|   <wbr>
|   "A"
